% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.r
\name{authenticate_inspera}
\alias{authenticate_inspera}
\title{Autentiser mot Inspera API}
\usage{
authenticate_inspera(api_key = NULL)
}
\arguments{
\item{api_key}{API-nøkkel for Inspera}
}
\value{
En liste som inneholder tilgangstoken og annen autentiseringsinformasjon
}
\description{
Autentiser mot Inspera API
}
