% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.r
\name{complete_inspera_contributor_workflow}
\alias{complete_inspera_contributor_workflow}
\title{Fullstendig arbeidsflyt for bidragsyterhåndtering}
\usage{
complete_inspera_contributor_workflow(
  source_test_id,
  target_test_id,
  committee_mapping = NULL,
  api_key = NULL
)
}
\arguments{
\item{source_test_id}{ID-en til kildetesten}

\item{target_test_id}{ID-en til måltesten}

\item{committee_mapping}{Manuell mapping av bidragsytere til komiteer}

\item{api_key}{API-nøkkel for Inspera}
}
\value{
En liste med resultater fra hele arbeidsflyten
}
\description{
Henter bidragsytere fra kildetest, konverterer til YAML med manuell komitétilordning,
og tilordner dem til en måltest.
}
