% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributor_management.r
\name{find_feide_username}
\alias{find_feide_username}
\title{Finn FEIDE-brukernavn fra en liste med bidragsytere}
\usage{
find_feide_username(contributors_list)
}
\arguments{
\item{contributors_list}{Liste med bidragsytere}
}
\value{
En liste med brukernavn og kilde for identifikasjonen
}
\description{
Finn FEIDE-brukernavn fra en liste med bidragsytere
}
