% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.r
\name{flatten_json}
\alias{flatten_json}
\title{Rekursiv funksjon for å flate ut JSON-objekter}
\usage{
flatten_json(json_obj, prefix = "")
}
\arguments{
\item{json_obj}{JSON-objekt som skal flates ut}

\item{prefix}{Prefiks for nøkler (brukes for nesting)}
}
\value{
En liste med flatet struktur
}
\description{
Rekursiv funksjon for å flate ut JSON-objekter
}
