% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_utils.r
\name{get_all_users}
\alias{get_all_users}
\title{Bestill og last ned fil med alle brukerdata fra Inspera}
\usage{
get_all_users(
  token = NULL,
  api_key = NULL,
  format = "csv",
  wait_time = NULL,
  max_attempts = NULL,
  save_path = NULL,
  debug = TRUE
)
}
\arguments{
\item{token}{Tilgangstoken (hvis ikke oppgitt, blir det hentet automatisk)}

\item{api_key}{API-nøkkel for Inspera (brukes kun hvis token ikke er oppgitt)}

\item{format}{Filformat for eksporten (default: "csv")}

\item{wait_time}{Hvor lenge skal vi vente mellom sjekkene på eksportstatus (sekunder)}

\item{max_attempts}{Maksimalt antall forsøk på å sjekke status}

\item{save_path}{Sti til hvor filen skal lagres (valgfri)}

\item{debug}{Om detaljert debug-informasjon skal skrives ut}
}
\value{
En data.frame med alle brukerdata
}
\description{
Bestill og last ned fil med alle brukerdata fra Inspera
}
