% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.r
\name{get_test_info}
\alias{get_test_info}
\title{Hent informasjon om en test fra Inspera API}
\usage{
get_test_info(test_id, token = NULL, api_key = NULL)
}
\arguments{
\item{test_id}{ID-en til testen}

\item{token}{Tilgangstoken (hvis ikke oppgitt, blir det hentet automatisk)}

\item{api_key}{API-nøkkel for Inspera (brukes kun hvis token ikke er oppgitt)}
}
\value{
En liste med testinformasjon
}
\description{
Hent informasjon om en test fra Inspera API
}
