% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_utils.r
\name{get_user_details}
\alias{get_user_details}
\title{Hent detaljert brukerinformasjon inkludert e-post}
\usage{
get_user_details(user_id, token)
}
\arguments{
\item{user_id}{Bruker-ID i Inspera}

\item{token}{Gyldig API-token}
}
\value{
Et brukerobjekt som inkluderer e-post hvis tilgjengelig
}
\description{
Hent detaljert brukerinformasjon inkludert e-post
}
