% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributor_management.r
\name{assign_contributors}
\alias{assign_contributors}
\title{Tilordne bidragsytere til en test med komiteer}
\usage{
assign_contributors(
  test_id,
  yaml_path,
  api_key = NULL,
  token = NULL,
  verbose = TRUE
)
}
\arguments{
\item{test_id}{ID-en til testen}

\item{yaml_path}{Sti til YAML-fil med bidragsytere}

\item{api_key}{API-nøkkel for Inspera}

\item{token}{Tilgangstoken (hvis ikke oppgitt, blir det hentet automatisk)}

\item{verbose}{Hvis TRUE, skriv ut detaljert informasjon om hver bidragsyter}
}
\value{
En liste med resultater fra tilordningen
}
\description{
Tilordne bidragsytere til en test med komiteer
}
