% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributor_management.r
\name{create_committees}
\alias{create_committees}
\title{Opprett komiteer i en test}
\usage{
create_committees(test_id, yaml_path, api_key = NULL)
}
\arguments{
\item{test_id}{ID-en til testen}

\item{yaml_path}{Sti til YAML-fil med komiténavn}

\item{api_key}{API-nøkkel for Inspera}
}
\value{
En liste med resultater for hver komité
}
\description{
Opprett komiteer i en test
}
