% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributor_management.r
\name{get_feide_contributors_and_save_csv}
\alias{get_feide_contributors_and_save_csv}
\title{Hent bidragsytere fra en test og lagre som CSV}
\usage{
get_feide_contributors_and_save_csv(test_id, api_key = NULL)
}
\arguments{
\item{test_id}{ID-en til testen}

\item{api_key}{API-nøkkel for Inspera}
}
\value{
En liste med dataframes for bidragsytere og annen informasjon
}
\description{
Denne funksjonen henter bidragsytere fra en Inspera-test, identifiserer
FEIDE-brukernavn og lagrer dataene i flere CSV-filer.
}
